Fix shared library (DLL) build on Windows.
- Now requires CMake 3.4, with the shiny new
  CMAKE_WINDOWS_EXPORT_ALL_SYMBOLS feature.
- Fixed typename definition in Array*DView.
- Fixed std::min/max bugs.
- Fixed float to int conversion warning in PerspectiveCamera and
  ProgressReporter.

+++ core/src/common/ProgressReporter.cpp
@@ -116,7 +116,7 @@ float ProgressReporter::approximateMillisecondsRemaining()
 
 float ProgressReporter::averageMillisecondsPerTask()
 {
-    return m_totalMillisecondsElapsed / m_nTasksCompleted;
+    return static_cast< float >( m_totalMillisecondsElapsed ) / m_nTasksCompleted;
 }
 
 void ProgressReporter::initialize( const std::string& prefix, int nTasks,
